feat: Add resource validation and layout optimization utilities
- Implemented `isResourceAffordable` function to check resource affordability based on player resources and card-specific storage.
- Created `analyzeCardLayout` function to analyze behaviors and generate a layout plan, including overflow handling.
- Developed `optimizeBehaviorsForSpace` function to reduce row count by converting resources to a more compact format.
- Added `detectTilePlacementScale` function to calculate scaling for tile placement icons based on behavior configurations.

refactor: Update imports for BehaviorSection component

- Removed explicit file extensions from imports of `BehaviorSection` across multiple components for consistency.

style: Enhance UI feedback for resource updates

- Introduced animations for value updates in the BottomResourceBar component to visually indicate changes in resources, actions, and effects.
- Added CSS animations for shining effects on text and icons to improve user experience.

chore: Add new CSS animations for UI enhancements

- Defined keyframes for value update shine, icon update shine, production shine, and button glow enhancements in the global CSS file.
