+++ frontend/src/components/ui/cards/BehaviorSection/constants.ts
@@ -1,0 +1,16 @@
+export const MAX_HORIZONTAL_ICONS = 7;
+export const MAX_CARD_ROWS = 4;
+
+export const CARD_STORAGE_RESOURCE_TYPES = [
+  "animals",
+  "microbes",
+  "floaters",
+  "science",
+  "asteroid",
+];
+
+export const TILE_PLACEMENT_TYPES = [
+  "city-placement",
+  "greenery-placement",
+  "ocean-placement",
+];
