+++ frontend/src/components/ui/cards/BehaviorSection/index.ts
@@ -1,0 +1,21 @@
+// Main component
+export { default } from "./BehaviorSection.tsx";
+
+// Types
+export type {
+  BehaviorSectionProps,
+  ClassifiedBehavior,
+  IconDisplayInfo,
+  LayoutPlan,
+  TileScaleInfo,
+  CardLayoutPlan,
+  IconContext,
+} from "./types.ts";
+
+// Constants
+export {
+  MAX_HORIZONTAL_ICONS,
+  MAX_CARD_ROWS,
+  CARD_STORAGE_RESOURCE_TYPES,
+  TILE_PLACEMENT_TYPES,
+} from "./constants.ts";
