+++ frontend/src/components/ui/cards/BehaviorSection/types.ts
@@ -1,0 +1,61 @@
+import { CardBehaviorDto, ResourcesDto } from "@/types/generated/api-types.ts";
+
+export interface BehaviorSectionProps {
+  behaviors?: CardBehaviorDto[];
+  playerResources?: ResourcesDto;
+  resourceStorage?: { [cardId: string]: number };
+  cardId?: string;
+  greyOutAll?: boolean;
+}
+
+export interface ClassifiedBehavior {
+  behavior: CardBehaviorDto;
+  type:
+    | "manual-action"
+    | "immediate-production"
+    | "immediate-effect"
+    | "triggered-effect"
+    | "auto-no-background"
+    | "discount";
+}
+
+export interface LayoutRequirement {
+  totalIcons: number;
+  separatorCount: number;
+  separatorPositions: number[];
+  behaviorType: string;
+  needsMultipleRows: boolean;
+  maxHorizontalIcons: number;
+}
+
+export interface IconDisplayInfo {
+  resourceType: string;
+  amount: number;
+  displayMode: "individual" | "number";
+  iconCount: number;
+}
+
+export interface LayoutPlan {
+  rows: IconDisplayInfo[][];
+  separators: Array<{ position: number; type: string }>;
+  totalRows: number;
+}
+
+export interface TileScaleInfo {
+  scale: 1 | 1.25 | 1.5 | 2;
+  tileType: string | null;
+  tileCount?: number;
+}
+
+export interface CardLayoutPlan {
+  behaviors: Array<{
+    behaviorIndex: number;
+    layoutPlan: LayoutPlan;
+    estimatedRows: number;
+  }>;
+  totalEstimatedRows: number;
+  needsOverflowHandling: boolean;
+  maxRows: number;
+}
+
+export type IconContext = "standalone" | "action" | "production" | "default";
