+++ frontend/src/components/ui/overlay/BottomResourceBar.tsx
@@ -30,6 +30,7 @@ interface BottomResourceBarProps {
   currentPlayer?: PlayerDto | null;
   gameState?: GameDto;
   playedCards?: CardDto[];
+  changedPaths?: Set<string>;
   onOpenCardEffectsModal?: () => void;
   onOpenCardsPlayedModal?: () => void;
   onOpenVictoryPointsModal?: () => void;
@@ -41,6 +42,7 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
   currentPlayer,
   gameState,
   playedCards = [],
+  changedPaths = new Set(),
   onOpenCardEffectsModal,
   onOpenCardsPlayedModal,
   onOpenVictoryPointsModal,
@@ -55,6 +57,11 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
   const effectsButtonRef = useRef<HTMLButtonElement>(null);
   const tagsButtonRef = useRef<HTMLButtonElement>(null);
   const storagesButtonRef = useRef<HTMLButtonElement>(null);
+
+  // Helper function to check if a path has changed
+  const hasPathChanged = (path: string): boolean => {
+    return changedPaths.has(path);
+  };
 
   // Map resource ID to ResourceType constant
   const getResourceType = (resourceId: string): string => {
@@ -209,49 +216,62 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
       {/* Resource Grid */}
       <div className="flex-[2] -translate-y-[30px] pointer-events-auto relative">
         <div className="grid grid-cols-6 gap-[15px] max-w-[500px]">
-          {playerResources.map((resource) => (
-            <div
-              key={resource.id}
-              className="flex flex-col items-center gap-1.5 bg-space-black-darker/90 border-2 rounded-xl p-2 transition-all duration-200 cursor-pointer relative overflow-hidden hover:-translate-y-0.5"
-              style={
-                {
-                  "--resource-color": resource.color,
-                  borderColor: resource.color,
-                  boxShadow: `0 0 10px ${resource.color}40`,
-                } as React.CSSProperties
-              }
-              onMouseEnter={(e) => {
-                e.currentTarget.style.boxShadow = `0 6px 20px rgba(0,0,0,0.4), 0 0 20px ${resource.color}`;
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.boxShadow = `0 0 10px ${resource.color}40`;
-              }}
-            >
-              <div className="inline-flex items-center justify-center bg-[linear-gradient(135deg,rgba(160,110,60,0.4)_0%,rgba(139,89,42,0.35)_100%)] border border-[rgba(160,110,60,0.5)] rounded px-2 py-1 shadow-[0_1px_3px_rgba(0,0,0,0.2)] mb-1 min-w-[28px]">
-                <span className="text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none">
-                  {resource.production}
-                </span>
-              </div>
+          {playerResources.map((resource) => {
+            const resourceChanged = hasPathChanged(
+              `currentPlayer.resources.${resource.id}`,
+            );
+            const productionChanged = hasPathChanged(
+              `currentPlayer.production.${resource.id}`,
+            );
 
-              {resource.id === "credits" ? (
-                <GameIcon
-                  iconType={ResourceTypeCredits}
-                  amount={resource.current}
-                  size="medium"
-                />
-              ) : (
-                <div className="flex items-center gap-1.5">
+            return (
+              <div
+                key={resource.id}
+                className="flex flex-col items-center gap-1.5 bg-space-black-darker/90 border-2 rounded-xl p-2 transition-all duration-200 cursor-pointer relative overflow-hidden hover:-translate-y-0.5"
+                style={
+                  {
+                    "--resource-color": resource.color,
+                    borderColor: resource.color,
+                    boxShadow: `0 0 10px ${resource.color}40`,
+                  } as React.CSSProperties
+                }
+                onMouseEnter={(e) => {
+                  e.currentTarget.style.boxShadow = `0 6px 20px rgba(0,0,0,0.4), 0 0 20px ${resource.color}`;
+                }}
+                onMouseLeave={(e) => {
+                  e.currentTarget.style.boxShadow = `0 0 10px ${resource.color}40`;
+                }}
+              >
+                <div className="inline-flex items-center justify-center bg-[linear-gradient(135deg,rgba(160,110,60,0.4)_0%,rgba(139,89,42,0.35)_100%)] border border-[rgba(160,110,60,0.5)] rounded px-2 py-1 shadow-[0_1px_3px_rgba(0,0,0,0.2)] mb-1 min-w-[28px]">
+                  <span
+                    className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${productionChanged ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
+                  >
+                    {resource.production}
+                  </span>
+                </div>
+
+                {resource.id === "credits" ? (
                   <GameIcon
-                    iconType={getResourceType(resource.id)}
+                    iconType={ResourceTypeCredits}
+                    amount={resource.current}
                     size="medium"
                   />
-                  <div className="text-lg font-bold text-white [text-shadow:0_1px_3px_rgba(0,0,0,0.8)]">
-                    {resource.current}
+                ) : (
+                  <div className="flex items-center gap-1.5">
+                    <GameIcon
+                      iconType={getResourceType(resource.id)}
+                      size="medium"
+                    />
+                    <div
+                      className={`text-lg font-bold text-white [text-shadow:0_1px_3px_rgba(0,0,0,0.8)] ${resourceChanged ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
+                    >
+                      {resource.current}
+                    </div>
                   </div>
-                </div>
-              )}
-            </div>
-          ))}
+                )}
+              </div>
+            );
+          })}
         </div>
       </div>
 
@@ -259,58 +279,27 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
       <div className="flex-1 flex items-center justify-end gap-3 -translate-y-[30px] pointer-events-auto relative">
         <button
           ref={actionsButtonRef}
-          className={`flex flex-col items-center gap-1 bg-space-black-darker/90 border-2 rounded-xl py-2.5 px-2 cursor-pointer transition-all duration-200 min-w-[60px] hover:-translate-y-0.5 ${
-            (currentPlayer?.actions?.length || 0) === 0
-              ? "border-[#969696] opacity-70 hover:opacity-80"
-              : (currentPlayer?.actions?.length || 0) <= 1
-                ? "border-[#ffc800]"
-                : "border-[#ff6464]"
-          }`}
-          style={{
-            boxShadow:
-              (currentPlayer?.actions?.length || 0) === 0
-                ? "0 0 10px #96969640"
-                : (currentPlayer?.actions?.length || 0) <= 1
-                  ? "0 0 10px #ffc80040"
-                  : "0 0 10px #ff646440",
-          }}
+          className="flex flex-col items-center gap-1 bg-space-black-darker/90 border-2 border-[#ff6464] rounded-xl py-2.5 px-2 cursor-pointer transition-all duration-200 min-w-[60px] hover:-translate-y-0.5"
+          style={{ boxShadow: "0 0 10px #ff646440" }}
           onMouseEnter={(e) => {
-            const color =
-              (currentPlayer?.actions?.length || 0) === 0
-                ? "#969696"
-                : (currentPlayer?.actions?.length || 0) <= 1
-                  ? "#ffc800"
-                  : "#ff6464";
-            e.currentTarget.style.boxShadow = `0 6px 20px rgba(0,0,0,0.4), 0 0 20px ${color}`;
+            e.currentTarget.style.boxShadow =
+              "0 6px 20px rgba(0,0,0,0.4), 0 0 20px #ff6464";
           }}
           onMouseLeave={(e) => {
-            const color =
-              (currentPlayer?.actions?.length || 0) === 0
-                ? "#969696"
-                : (currentPlayer?.actions?.length || 0) <= 1
-                  ? "#ffc800"
-                  : "#ff6464";
-            e.currentTarget.style.boxShadow = `0 0 10px ${color}40`;
+            e.currentTarget.style.boxShadow = "0 0 10px #ff646440";
           }}
           onClick={handleOpenActionsPopover}
         >
           <div
             className="text-base font-bold [filter:drop-shadow(0_1px_2px_rgba(0,0,0,0.5))] flex items-center gap-[2px] h-[32px] w-[32px] justify-center"
-            style={{
-              color:
-                (currentPlayer?.actions?.length || 0) === 0
-                  ? "#969696"
-                  : (currentPlayer?.actions?.length || 0) <= 1
-                    ? "#ffc800"
-                    : "#ff6464",
-            }}
+            style={{ color: "#ff6464" }}
           >
             <span className="text-[8px] leading-none translate-y-[1px]">●</span>
             <span className="text-[8px] leading-none translate-y-[1px]">●</span>
             <span className="text-[23px] leading-none">→</span>
           </div>
           <div
-            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${(currentPlayer?.actions?.length || 0) === 0 ? "text-white/60" : ""}`}
+            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${hasPathChanged("currentPlayer.actions") ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
           >
             {currentPlayer?.actions?.length || 0}
           </div>
@@ -342,7 +331,9 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
               <span className="text-[10px] leading-none">●</span>
             </div>
           </div>
-          <div className="text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none">
+          <div
+            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${hasPathChanged("currentPlayer.effects") ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
+          >
             {currentPlayer?.effects?.length || 0}
           </div>
           <div className="text-[10px] font-medium text-white/90 uppercase tracking-[0.5px] [text-shadow:0_1px_2px_rgba(0,0,0,0.8)]">
@@ -374,7 +365,9 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
               <span>●</span>
             </div>
           </div>
-          <div className="text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none">
+          <div
+            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${hasPathChanged("currentPlayer.playedCards") ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
+          >
             {tagCounts.reduce((sum, tag) => sum + tag.count, 0)}
           </div>
           <div className="text-[10px] font-medium text-white/90 uppercase tracking-[0.5px] [text-shadow:0_1px_2px_rgba(0,0,0,0.8)]">
@@ -406,7 +399,9 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
               <span>●</span>
             </div>
           </div>
-          <div className="text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none">
+          <div
+            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${hasPathChanged("currentPlayer.resourceStorage") ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
+          >
             {storageCardsCount}
           </div>
           <div className="text-[10px] font-medium text-white/90 uppercase tracking-[0.5px] [text-shadow:0_1px_2px_rgba(0,0,0,0.8)]">
@@ -432,7 +427,9 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
           >
             ↓
           </div>
-          <div className="text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none">
+          <div
+            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${hasPathChanged("currentPlayer.playedCards") ? "[animation:valueUpdateShine_0.8s_ease-in-out]" : ""}`}
+          >
             {playedCardsCount}
           </div>
           <div className="text-[10px] font-medium text-white/90 uppercase tracking-[0.5px] [text-shadow:0_1px_2px_rgba(0,0,0,0.8)]">
@@ -459,7 +456,14 @@ const BottomResourceBar: React.FC<BottomResourceBarProps> = ({
             <span className="text-3xl absolute">○</span>
             <span className="text-lg absolute">●</span>
           </div>
-          <div className="text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none">
+          <div
+            className={`text-sm font-bold text-white [text-shadow:0_1px_2px_rgba(0,0,0,0.8)] leading-none ${
+              hasPathChanged("currentPlayer.victoryPoints") ||
+              hasPathChanged("currentPlayer.terraformRating")
+                ? "[animation:valueUpdateShine_0.8s_ease-in-out]"
+                : ""
+            }`}
+          >
             {currentPlayer?.victoryPoints || 0}
           </div>
           <div className="text-[10px] font-medium text-white/90 uppercase tracking-[0.5px] [text-shadow:0_1px_2px_rgba(0,0,0,0.8)]">
