+++ frontend/src/components/ui/popover/ActionsPopover.tsx
@@ -3,7 +3,7 @@ import {
   PlayerActionDto,
   GameDto,
 } from "../../../types/generated/api-types.ts";
-import BehaviorSection from "../cards/BehaviorSection.tsx";
+import BehaviorSection from "../cards/BehaviorSection";
 import {
   canPerformActions,
   hasActionsAvailable,
